feat: move pingpong in dedicated project
