+++ settings.gradle.kts
@@ -18,4 +18,5 @@ include(":jakta-dsl")
 include(":jakta-distributed:jakta-distributed-client")
 include(":jakta-distributed:jakta-distributed-broker")
 include(":jakta-distributed:jakta-distributed-common")
-include("jakta-distributed:jakta-distributed-adapter")
+include(":jakta-distributed:jakta-distributed-adapter")
+include(":jakta-distributed:jakta-distributed-examples")
